requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
aiosqlite==0.21.0
python-dotenv==1.1.0
bcrypt==4.3.0
itsdangerous==2.2.0
jinja2==3.1.6
python-multipart==0.0.20
greenlet==3.4.0
openai==1.75.0
httpx==0.28.1
